requirements.txt
Django==3.2.9
django_ipware==4.0.0
folium==0.12.1
geocoder==1.38.1
geopy==2.2.0
